chore: moving some sources to core folder
