+++ srcs/core/errors/error_handling.c
@@ -1,4 +1,4 @@
-#include "../../includes/error_handling.h"
+#include "../../../includes/error_handling.h"
 #include <errno.h>
 #include <stdarg.h>
 #include <syslog.h>
