+++ srcs/core/handlers/methods/get_handler.c
