+++ srcs/core/handlers/methods/unsupported_handler.c
