+++ srcs/core/handlers/threads/client_theads.c
