+++ srcs/core/router/router.c
@@ -1,7 +1,7 @@
-#include "../../includes/http.h"
-#include "../../includes/router.h"
-#include "../../includes/httpc.h"
-#include "../../includes/structs/route.h"
+#include "../../../includes/http.h"
+#include "../../../includes/router.h"
+#include "../../../includes/httpc.h"
+#include "../../../includes/structs/route.h"
 #include <stdbool.h>
 #include <errno.h>
